Extract vfr intensity to file
